+++ src/sdk/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Support for getting, setting and clearing of (folder based) column defaults #949 [jansenbe - Bert Jansen]
 - Requesting re-indexing of a web or list #950 [jansenbe - Bert Jansen]
 - RateLimit header processing capability, turned off by default [jansenbe - Bert Jansen]
+- Added `DeleteRecycledSiteCollection` methods to delete site collections from the recycle bin [jansenbe - Bert Jansen]
 
 ### Changed